+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 - Add account checking in `/etc/passwd`
 - Add filesystem scanning
 - Add checking of ingress nginx
+- Add BearerToken for authentication
+- Add insecure and server flags in k8s analysis
 
 ## improvements
 - Add the counter of each severity 
 - Add some rules of annotation checking
+- Delete the inside flag due to duplicate
+- Add `.dockerconfigjson` in secret checking
 
 ## improvements
 - Add the date of kernel compiling checking in checking of kernel version