+++ CHANGELOG.md
@@ -13,10 +13,12 @@
 - Delete the inside flag due to duplicate
 - Add `.dockerconfigjson` in secret checking
 - Add Docker Histories environment checking
-
-## improvements
 - Add the date of kernel compiling checking in checking of kernel version
 - Add the error output in image saving
+
+## fixed
+- Fix the out of range in container extract
+
 
 # 1.0.8 (2023.6.6)
 ## features